# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema Escolar - API
  description: API para gestión de computadoras y reservas en laboratorios escolares.
  version: 1.0.0
servers:
  - url: http://localhost:5051
    description: Desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/computers:
    get:
      summary: Listar computadoras
      tags: [Computers]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: status
          in: query
          schema: { type: string, enum: [available, in_use, maintenance, out_of_service] }
      responses:
        '200':
          description: Lista paginada de computadoras
        '401':
          description: No autenticado
    post:
      summary: Crear computadora (admin/teacher)
      tags: [Computers]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [code, description]
              properties:
                code: { type: string }
                description: { type: string }
                serial_number: { type: string, nullable: true }
                hardware_id: { type: string, nullable: true }
      responses:
        '201':
          description: Computadora creada
        '400':
          description: Validación fallida
        '403':
          description: Rol insuficiente
...
  /api/reservations:
    get:
      summary: Listar reservas
      tags: [Reservations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string, enum: [active, completed, cancelled] }
        - name: user_id
          in: query
          schema: { type: integer }
        - name: computer_id
          in: query
          schema: { type: integer }
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Lista paginada de reservas
    post:
      summary: Crear reserva
      tags: [Reservations]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [computer_id, start_time, end_time]
              properties:
                computer_id: { type: integer }
                start_time: { type: string, format: date-time }
                end_time: { type: string, format: date-time }
      responses:
        '201':
          description: Reserva creada
        '400':
          description: Validación o conflicto de horario
  /api/users:
    get:
      summary: Listar usuarios (solo admin)
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: role
          in: query
          schema: { type: string, enum: [admin, teacher, student] }
      responses:
        '200':
          description: Lista paginada de usuarios
        '403':
          description: Solo admin
...